Bugfix: int8 diameter was blocking changed to int16
Bugfix: int8 diameter was blocking changed to int16

+++ saveCaseDriver.go
@@ -98,7 +98,7 @@ func saveCaseDriver(r *http.Request, ctx appengine.Context) (string) {
 		FrequencyLow:		int32(driverFrequencyLow), // int32
 		FrequencyHigh:		int32(driverFrequencyHigh), // int32
 		
-		Diameter:			int8(driverdiameter), // int8
+		Diameter:			int16(driverdiameter), // int16
 		Price:				int32(driverprice), // int32
 		
 		BlobKey:			blobkey,
